fix warnings and appveyor again
Former-commit-id: ba075fb187aaefe32d332d71b651aa5ed1ca8379

+++ source/fir/Types/SingleTypes.cpp
@@ -151,7 +151,7 @@ namespace fir
 	static std::vector<std::unordered_map<std::string, PolyPlaceholderType*>> cache;
 	PolyPlaceholderType* PolyPlaceholderType::get(const std::string& n, int group)
 	{
-		while(group >= cache.size())
+		while((size_t) group >= cache.size())
 			cache.push_back({ });
 
 		if(auto it = cache[group].find(n); it != cache[group].end())
